Gcc compile fixes.
git-svn-id: http://vtp.googlecode.com/svn/trunk@7818 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/RoadMap.cpp
@@ -260,8 +260,8 @@ void TNode::AdjustForLights()
 						newAngle = -newAngle;
 					}
 					printf("%i:%f, %i:%f, %f, %f",
-						i, m_fLinkAngle[i],
-						j, m_fLinkAngle[j],
+						(int) i, m_fLinkAngle[i],
+						(int) j, m_fLinkAngle[j],
 						newAngle, bestAngle);
 					if (newAngle < bestAngle) {
 						bestChoiceA = i;
